// packages/dnb-eufemia/src/extensions/forms/blocks/ChildrenWithAge/ChildrenWithAgeTranslations.d.ts
export declare const translations: {
    'nb-NO': {
        ChildrenWithAge: {
            hasChildren: {
                title: string;
                fieldLabel: string;
                required: string;
            };
            countChildren: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            childrenAge: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            hasJointResponsibility: {
                fieldLabel: string;
                required: string;
            };
            jointResponsibilityExpenses: {
                fieldLabel: string;
                required: string;
            };
            usesDaycare: {
                fieldLabel: string;
                required: string;
                helpText: string;
            };
            dayCareExpenses: {
                fieldLabel: string;
                required: string;
            };
        };
    };
    'sv-SE': {
        ChildrenWithAge: {
            hasChildren: {
                title: string;
                fieldLabel: string;
                required: string;
            };
            countChildren: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            childrenAge: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            hasJointResponsibility: {
                fieldLabel: string;
                required: string;
            };
            jointResponsibilityExpenses: {
                fieldLabel: string;
                required: string;
            };
            usesDaycare: {
                fieldLabel: string;
                required: string;
                helpText: string;
            };
            dayCareExpenses: {
                fieldLabel: string;
                required: string;
            };
        };
    };
    'da-DK': {
        ChildrenWithAge: {
            hasChildren: {
                title: string;
                fieldLabel: string;
                required: string;
            };
            countChildren: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            childrenAge: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            hasJointResponsibility: {
                fieldLabel: string;
                required: string;
            };
            jointResponsibilityExpenses: {
                fieldLabel: string;
                required: string;
            };
            usesDaycare: {
                fieldLabel: string;
                required: string;
                helpText: string;
            };
            dayCareExpenses: {
                fieldLabel: string;
                required: string;
            };
        };
    };
    'en-GB': {
        ChildrenWithAge: {
            hasChildren: {
                title: string;
                fieldLabel: string;
                required: string;
            };
            countChildren: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            childrenAge: {
                fieldLabel: string;
                required: string;
                suffix: string;
            };
            hasJointResponsibility: {
                fieldLabel: string;
                required: string;
            };
            jointResponsibilityExpenses: {
                fieldLabel: string;
                required: string;
            };
            usesDaycare: {
                fieldLabel: string;
                required: string;
                helpText: string;
            };
            dayCareExpenses: {
                fieldLabel: string;
                required: string;
            };
        };
    };
};
export type Translation = (typeof translations)[keyof typeof translations];